"combined "preZentra.py" into "zentra.py", and create a function for the database writing."
